# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VertiGrow Vertical Farming Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ZoneStatus = (typeof ZoneStatus)[keyof typeof ZoneStatus];

export const ZoneStatus = {
  active: "active",
  idle: "idle",
  maintenance: "maintenance",
} as const;

export interface Zone {
  id: number;
  name: string;
  cropType: string;
  rackCount: number;
  status: ZoneStatus;
  createdAt: string;
  updatedAt: string;
}

export type CreateZoneBodyStatus =
  (typeof CreateZoneBodyStatus)[keyof typeof CreateZoneBodyStatus];

export const CreateZoneBodyStatus = {
  active: "active",
  idle: "idle",
  maintenance: "maintenance",
} as const;

export interface CreateZoneBody {
  name: string;
  cropType: string;
  rackCount: number;
  status?: CreateZoneBodyStatus;
}

export type UpdateZoneBodyStatus =
  (typeof UpdateZoneBodyStatus)[keyof typeof UpdateZoneBodyStatus];

export const UpdateZoneBodyStatus = {
  active: "active",
  idle: "idle",
  maintenance: "maintenance",
} as const;

export interface UpdateZoneBody {
  name?: string;
  cropType?: string;
  rackCount?: number;
  status?: UpdateZoneBodyStatus;
}

export interface SensorReading {
  id: number;
  zoneId: number;
  /** Temperature in Celsius */
  temperature: number;
  /** Relative humidity percentage */
  humidity: number;
  /** Soil moisture percentage */
  soilMoisture: number;
  /** Water pH level */
  phLevel: number;
  recordedAt: string;
}

export interface CreateSensorReadingBody {
  zoneId: number;
  temperature: number;
  humidity: number;
  soilMoisture: number;
  phLevel: number;
}

export type ControlType = (typeof ControlType)[keyof typeof ControlType];

export const ControlType = {
  led_lighting: "led_lighting",
  cooling_fan: "cooling_fan",
  water_pump: "water_pump",
  nutrient_pump: "nutrient_pump",
} as const;

export interface Control {
  id: number;
  zoneId: number;
  type: ControlType;
  isActive: boolean;
  /**
   * Intensity percentage (0-100), null if not applicable
   * @nullable
   */
  intensity?: number | null;
  updatedAt: string;
}

export interface UpdateControlBody {
  isActive?: boolean;
  /** @nullable */
  intensity?: number | null;
}

export type AlertSeverity = (typeof AlertSeverity)[keyof typeof AlertSeverity];

export const AlertSeverity = {
  critical: "critical",
  warning: "warning",
  info: "info",
} as const;

export type AlertType = (typeof AlertType)[keyof typeof AlertType];

export const AlertType = {
  temperature_high: "temperature_high",
  temperature_low: "temperature_low",
  humidity_high: "humidity_high",
  humidity_low: "humidity_low",
  ph_out_of_range: "ph_out_of_range",
  soil_dry: "soil_dry",
  reservoir_low: "reservoir_low",
  system_offline: "system_offline",
} as const;

export type AlertStatus = (typeof AlertStatus)[keyof typeof AlertStatus];

export const AlertStatus = {
  active: "active",
  acknowledged: "acknowledged",
  resolved: "resolved",
} as const;

export interface Alert {
  id: number;
  /** @nullable */
  zoneId?: number | null;
  severity: AlertSeverity;
  type: AlertType;
  message: string;
  status: AlertStatus;
  createdAt: string;
  updatedAt: string;
}

export type RecommendationCategory =
  (typeof RecommendationCategory)[keyof typeof RecommendationCategory];

export const RecommendationCategory = {
  harvest_timing: "harvest_timing",
  nutrient_adjustment: "nutrient_adjustment",
  lighting_schedule: "lighting_schedule",
  watering_cycle: "watering_cycle",
  environmental_control: "environmental_control",
} as const;

export type RecommendationPriority =
  (typeof RecommendationPriority)[keyof typeof RecommendationPriority];

export const RecommendationPriority = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface Recommendation {
  id: number;
  /** @nullable */
  zoneId?: number | null;
  category: RecommendationCategory;
  title: string;
  description: string;
  priority: RecommendationPriority;
  createdAt: string;
}

export interface DashboardSummary {
  totalZones: number;
  activeZones: number;
  activeAlerts: number;
  criticalAlerts: number;
  avgTemperature: number;
  avgHumidity: number;
  avgPhLevel: number;
  avgSoilMoisture: number;
  activeControls: number;
  pendingRecommendations: number;
}

export type GetSensorHistoryParams = {
  zoneId: number;
  metric?: GetSensorHistoryMetric;
  limit?: number;
};

export type GetSensorHistoryMetric =
  (typeof GetSensorHistoryMetric)[keyof typeof GetSensorHistoryMetric];

export const GetSensorHistoryMetric = {
  temperature: "temperature",
  humidity: "humidity",
  soilMoisture: "soilMoisture",
  phLevel: "phLevel",
} as const;

export type ListAlertsParams = {
  status?: ListAlertsStatus;
};

export type ListAlertsStatus =
  (typeof ListAlertsStatus)[keyof typeof ListAlertsStatus];

export const ListAlertsStatus = {
  active: "active",
  acknowledged: "acknowledged",
  resolved: "resolved",
} as const;