# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VertiGrow Vertical Farming Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all farm zones
 */
export const ListZonesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  cropType: zod.string(),
  rackCount: zod.number(),
  status: zod.enum(["active", "idle", "maintenance"]),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListZonesResponse = zod.array(ListZonesResponseItem);

/**
 * @summary Create a new farm zone
 */
export const CreateZoneBody = zod.object({
  name: zod.string(),
  cropType: zod.string(),
  rackCount: zod.number(),
  status: zod.enum(["active", "idle", "maintenance"]).optional(),
});

/**
 * @summary Get a zone by ID
 */
export const GetZoneParams = zod.object({
  id: zod.coerce.number(),
});

export const GetZoneResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  cropType: zod.string(),
  rackCount: zod.number(),
  status: zod.enum(["active", "idle", "maintenance"]),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update a zone
 */
export const UpdateZoneParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateZoneBody = zod.object({
  name: zod.string().optional(),
  cropType: zod.string().optional(),
  rackCount: zod.number().optional(),
  status: zod.enum(["active", "idle", "maintenance"]).optional(),
});

export const UpdateZoneResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  cropType: zod.string(),
  rackCount: zod.number(),
  status: zod.enum(["active", "idle", "maintenance"]),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a zone
 */
export const DeleteZoneParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get latest sensor readings for all zones
 */
export const ListSensorsResponseItem = zod.object({
  id: zod.number(),
  zoneId: zod.number(),
  temperature: zod.number().describe("Temperature in Celsius"),
  humidity: zod.number().describe("Relative humidity percentage"),
  soilMoisture: zod.number().describe("Soil moisture percentage"),
  phLevel: zod.number().describe("Water pH level"),
  recordedAt: zod.coerce.date(),
});
export const ListSensorsResponse = zod.array(ListSensorsResponseItem);

/**
 * @summary Ingest a new sensor reading (simulated IoT input)
 */
export const CreateSensorReadingBody = zod.object({
  zoneId: zod.number(),
  temperature: zod.number(),
  humidity: zod.number(),
  soilMoisture: zod.number(),
  phLevel: zod.number(),
});

/**
 * @summary Get historical sensor readings for a zone
 */
export const getSensorHistoryQueryLimitDefault = 48;

export const GetSensorHistoryQueryParams = zod.object({
  zoneId: zod.coerce.number(),
  metric: zod
    .enum(["temperature", "humidity", "soilMoisture", "phLevel"])
    .optional(),
  limit: zod.coerce.number().default(getSensorHistoryQueryLimitDefault),
});

export const GetSensorHistoryResponseItem = zod.object({
  id: zod.number(),
  zoneId: zod.number(),
  temperature: zod.number().describe("Temperature in Celsius"),
  humidity: zod.number().describe("Relative humidity percentage"),
  soilMoisture: zod.number().describe("Soil moisture percentage"),
  phLevel: zod.number().describe("Water pH level"),
  recordedAt: zod.coerce.date(),
});
export const GetSensorHistoryResponse = zod.array(GetSensorHistoryResponseItem);

/**
 * @summary List all control system states
 */
export const ListControlsResponseItem = zod.object({
  id: zod.number(),
  zoneId: zod.number(),
  type: zod.enum([
    "led_lighting",
    "cooling_fan",
    "water_pump",
    "nutrient_pump",
  ]),
  isActive: zod.boolean(),
  intensity: zod
    .number()
    .nullish()
    .describe("Intensity percentage (0-100), null if not applicable"),
  updatedAt: zod.coerce.date(),
});
export const ListControlsResponse = zod.array(ListControlsResponseItem);

/**
 * @summary Update a control state (toggle LED, fan, pump)
 */
export const UpdateControlParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateControlBody = zod.object({
  isActive: zod.boolean().optional(),
  intensity: zod.number().nullish(),
});

export const UpdateControlResponse = zod.object({
  id: zod.number(),
  zoneId: zod.number(),
  type: zod.enum([
    "led_lighting",
    "cooling_fan",
    "water_pump",
    "nutrient_pump",
  ]),
  isActive: zod.boolean(),
  intensity: zod
    .number()
    .nullish()
    .describe("Intensity percentage (0-100), null if not applicable"),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary List alerts (anomalies and warnings)
 */
export const ListAlertsQueryParams = zod.object({
  status: zod.enum(["active", "acknowledged", "resolved"]).optional(),
});

export const ListAlertsResponseItem = zod.object({
  id: zod.number(),
  zoneId: zod.number().nullish(),
  severity: zod.enum(["critical", "warning", "info"]),
  type: zod.enum([
    "temperature_high",
    "temperature_low",
    "humidity_high",
    "humidity_low",
    "ph_out_of_range",
    "soil_dry",
    "reservoir_low",
    "system_offline",
  ]),
  message: zod.string(),
  status: zod.enum(["active", "acknowledged", "resolved"]),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListAlertsResponse = zod.array(ListAlertsResponseItem);

/**
 * @summary Acknowledge an alert
 */
export const AcknowledgeAlertParams = zod.object({
  id: zod.coerce.number(),
});

export const AcknowledgeAlertResponse = zod.object({
  id: zod.number(),
  zoneId: zod.number().nullish(),
  severity: zod.enum(["critical", "warning", "info"]),
  type: zod.enum([
    "temperature_high",
    "temperature_low",
    "humidity_high",
    "humidity_low",
    "ph_out_of_range",
    "soil_dry",
    "reservoir_low",
    "system_offline",
  ]),
  message: zod.string(),
  status: zod.enum(["active", "acknowledged", "resolved"]),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary List AI-driven crop recommendations
 */
export const ListRecommendationsResponseItem = zod.object({
  id: zod.number(),
  zoneId: zod.number().nullish(),
  category: zod.enum([
    "harvest_timing",
    "nutrient_adjustment",
    "lighting_schedule",
    "watering_cycle",
    "environmental_control",
  ]),
  title: zod.string(),
  description: zod.string(),
  priority: zod.enum(["high", "medium", "low"]),
  createdAt: zod.coerce.date(),
});
export const ListRecommendationsResponse = zod.array(
  ListRecommendationsResponseItem,
);

/**
 * @summary Get aggregated dashboard summary stats
 */
export const GetDashboardSummaryResponse = zod.object({
  totalZones: zod.number(),
  activeZones: zod.number(),
  activeAlerts: zod.number(),
  criticalAlerts: zod.number(),
  avgTemperature: zod.number(),
  avgHumidity: zod.number(),
  avgPhLevel: zod.number(),
  avgSoilMoisture: zod.number(),
  activeControls: zod.number(),
  pendingRecommendations: zod.number(),
});